added the basic method GetByConditionsAsync to the repository interface -> use the predicate in the repository->  get entities
