+++ OnlineStore.BLL/Services/Implementations/UserService.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace OnlineStore.BLL.Services
+namespace OnlineStore.BLL.Services.Implementations
 {
-    internal class IOrderItemService
+    internal class UserService
     {
     }
 }
