+++ OnlineStore.BLL/Services/Interfaces/IOrderService.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace OnlineStore.BLL.Services
+namespace OnlineStore.BLL.Services.Interfaces
 {
     internal class IOrderService
     {
