+++ OnlineStore.DAL/Repositories/Implementations/OrderRepository.cs
@@ -1,10 +1,11 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using OnlineStore.DAL.Entities;
+using OnlineStore.DAL.Repositories.Interfaces;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace OnlineStore.DAL.Repositories
+namespace OnlineStore.DAL.Repositories.Implementations
 {
     public class OrderRepository : Repository<Order>, IOrderRepository
     {
@@ -14,7 +15,7 @@ namespace OnlineStore.DAL.Repositories
 
         public async Task<IEnumerable<Order>> GetOrdersByUserIdAsync(int userId)
         {
-            return await _context.Orders.Where(o => o.UserId == userId).ToListAsync();
+            return await GetByConditionAsync(order => order.UserId == userId);
         }
     }
 }
