+++ OnlineStore.DAL/Repositories/Interfaces/ICategoryRepository.cs
@@ -1,6 +1,6 @@
 ﻿using OnlineStore.DAL.Entities;
 
-namespace OnlineStore.DAL.Repositories
+namespace OnlineStore.DAL.Repositories.Interfaces
 {
     public interface ICategoryRepository : IRepository<Category>
     {
