+++ OnlineStore.DAL/Repositories/Interfaces/IProductRepository.cs
@@ -1,6 +1,6 @@
 ﻿using OnlineStore.DAL.Entities;
 
-namespace OnlineStore.DAL.Repositories
+namespace OnlineStore.DAL.Repositories.Interfaces
 {
     public interface IProductRepository : IRepository<Product>
     {
