+++ OnlineStore.DAL/Repositories/Interfaces/IRepository.cs
@@ -1,8 +1,9 @@
 ﻿using System.Collections.Generic;
+using System.Linq.Expressions;
 using System.Threading.Tasks;
 
 
-namespace OnlineStore.DAL.Repositories
+namespace OnlineStore.DAL.Repositories.Interfaces
 {
     public interface IRepository<T> where T : class
     {
@@ -11,6 +12,8 @@ namespace OnlineStore.DAL.Repositories
         Task AddAsync(T entity);
         Task UpdateAsync(T entity);
         Task DeleteAsync(int id);
+
+        Task<IEnumerable<T>> GetByConditionAsync(Expression<Func<T, bool>> predicate);
     }
 
 }
