--- OnlineStore.DAL/Repositories/ProductRepository.cs
@@ -1,20 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using OnlineStore.DAL.Entities;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace OnlineStore.DAL.Repositories
-{
-    public class ProductRepository : Repository<Product>, IProductRepository
-    {
-        public ProductRepository(ApplicationDbContext context) : base(context)
-        {
-        }
-
-        public async Task<IEnumerable<Product>> GetProductsByCategoryAsync(int categoryId)
-        {
-            return await _context.Products.Where(p => p.CategoryId == categoryId).ToListAsync();
-        }
-    }
-}
